Now works with non Winbond modules (beta) by taking the flash memory size in bits as an argument in flash.begin('CHIPSIZE');
