+++ defines.h
@@ -48,6 +48,41 @@
 #define BLOCK64ERASE 0xD8
 
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
+//                     General size definitions                       //
+//            B = Bytes; KB = Kilo bits; MB = Mega bits               //
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
+#define B1            1L
+#define B2            2L
+#define B4            4L
+#define B8            8L
+#define B16           16L
+#define B32           32L
+#define B64           64L
+#define B80           80L
+#define B128          128L
+#define B256          256L
+#define B512          512L
+#define KB1           B1 * K
+#define KB2           B2 * K
+#define KB4           B4 * K
+#define KB8           B8 * K
+#define KB16          B16 * K
+#define KB32          B32 * K
+#define KB64          B64 * K
+#define KB128         B128 * K
+#define KB256         B256 * K
+#define KB512         B512 * K
+#define MB1           B1 * M
+#define MB2           B2 * M
+#define MB4           B4 * M
+#define MB8           B8 * M
+#define MB16          B16 * M
+#define MB32          B32 * M
+#define MB64          B64 * M
+#define MB128         B128 * M
+#define MB256         B256 * M
+#define MB512         B512 * M
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
 //					Chip specific instructions 						  //
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
 
@@ -83,6 +118,9 @@
 #define arrayLen(x)   (sizeof(x) / sizeof(*x))
 #define lengthOf(x)   (sizeof(x))/sizeof(byte)
 #define maxAddress    capacity
+#define K             1024L
+#define M             K * K
+#define S             1000L
 
 #if defined (ARDUINO_ARCH_ESP8266)
 #define CS 15
