+++ examples/FlashDiagnostics/FlashDiagnostic_functions.ino
