+++ examples/FlashDiagnostics/FlashDiagnostics.ino
@@ -17,6 +17,9 @@
 
 #include<SPIFlash.h>
 
+//Define a flash memory size (if using non-Winbond memory) according to the list in defines.h
+#define CHIPSIZE MB8
+
 #if defined(ARDUINO_SAMD_ZERO) && defined(SERIAL_PORT_USBVIRTUAL)
 // Required for Serial on Zero based boards
 #define Serial SERIAL_PORT_USBVIRTUAL
@@ -43,7 +46,7 @@ void setup() {
     Serial.print(F("."));
   }
   Serial.println();
-  flash.begin();
+  flash.begin(CHIPSIZE); //use flash.begin(CHIPSIZE) if using non-Winbond flash (Refer to '#define CHIPSIZE' above
   Serial.println();
   Serial.println();
